refactor: simplify main function by delegating command handling to cli module
