+++ src/keystore/tests.rs
@@ -91,4 +91,4 @@ fn test_hsm_backend_not_implemented() {
         .unwrap_err()
         .to_string()
         .contains("not yet implemented"));
-} 
+}
