+++ src/security/mod.rs
@@ -3,4 +3,4 @@ pub mod validator;
 #[cfg(test)]
 mod tests;
 
-pub use validator::SecurityValidator; 
+pub use validator::SecurityValidator;
