+++ src/security/validator.rs
@@ -99,4 +99,4 @@ impl SecurityValidator {
             "Chain ID {chain_id} is not allowed"
         )))
     }
-} 
+}
